fix(shared-ui): unify crossOriginHref resolver — preserve path + query
Studio and shared-ui had diverging resolvers: studioCrossOriginHref
preserved item.path but dropped item.query, while defaultCrossOriginHref
preserved query but forced '/' (dropping path). A nav item with both
(e.g. Canvas plugin=galaxy at '/') lost data in one direction or the
other — clicking it from different origins produced different URLs.

Consolidate into one resolver in shared-ui that routes every bounce
through a buildHref helper: append ?host= first, then merge item.query.
Studio→vector bounce for playground/compare paths stays inside the
default (exported as isVectorPath for apps that need to reuse it).
Delete studioCrossOriginHref — studio now uses the default.

Verify:
- from feed.*, Canvas (path='/', query={plugin:'galaxy'}, studio=canvas.*)
  → https://canvas.*/?host=feed.*&plugin=galaxy
- from vector.*, Memory (path='/map', no studio) → studio.*/map?host=vector.*
- from studio.*, Compare (path='/compare', no studio) → vector.*/compare?host=studio.*
